+++ files/chef-marketplace-cookbooks/chef-marketplace/files/default/chef-marketplace-plugin-config.rb
@@ -3,7 +3,12 @@ gen_api_fqdn = lambda do
     marketplace = JSON.parse(IO.read('/etc/chef-marketplace/chef-marketplace-running.json'))['chef-marketplace']
 
     api_fqdn marketplace['api_fqdn']
-    license['nodes'] = marketplace['license_count']
+    license['nodes'] =
+      if marketplace['reckoner']['enabled']
+        999_999_999
+      else
+        marketplace['license_count']
+      end
 
     if marketplace['role'] == 'aio'
       redirect_uri = "https://#{marketplace['api_fqdn']}:#{marketplace['analytics']['ssl_port']}"